requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib
python-multipart
pandas
openpyxl
python-dotenv
passlib[bcrypt]
psycopg2-binary
jinja2
pydantic[email]
fastapi-mail